src/juror_server/requirements.txt
fastapi~=0.120.0
pyyaml~=6.0.2
opencv-python~=4.12.0.88
numpy~=2.2.6
phx-class-registry~=5.1.1
matplotlib~=3.10.6
transformers~=4.56.2
pillow~=11.3.0
phx-class-registry~=5.1.1
pydantic~=2.11.9
compress_pickle~=2.1.0
mlflow~=3.5.1


# Torch and torchvision installation based on platform and CUDA version
# for CUDA 12.9 torch on Windows
torch --index-url https://download.pytorch.org/whl/cu129; sys_platform == 'Windows'
torchvision --index-url https://download.pytorch.org/whl/cu129; sys_platform == 'Windows'

# Torch und torchvision für Linux mit CUDA 12.9
torch --index-url https://download.pytorch.org/whl/cu129; sys_platform == 'linux'
torchvision --index-url https://download.pytorch.org/whl/cu129; sys_platform == 'linux'

# for MPS accelerated torch on Mac
torch; sys_platform == 'darwin'
torchvision; sys_platform == 'darwin'

fiftyone~=1.8.0